% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMBstats.r
\name{getMBstats}
\alias{getMBstats}
\title{Calculate statistics for the second-tier projection}
\usage{
getMBstats(seu, group.by)
}
\arguments{
\item{seu}{Modified query Seurat object from \code{mapToMB()}.}

\item{group.by}{Name of one metadata column to group cells by.}
}
\value{
A matrix of 154 rows. Each column corresponds to one group specified by group.by. 
Row 1-17 detail midbrain celltype proportion
Row 18-23 detail midbrain progenitor proportion
Row 24-31 detail midbrain neuron proportion
Row 32-131 detail ventral score
Row 132-137 detail the correlation between midbrain celltype proportion of current query dataset and that reported by La Manno 2016
Row 138-152 detail midbrain celltype identity NES
}
\description{
Calculate statistics for the second-tier projection
}
\author{
Lisheng Xu
}
