% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWBstats.r
\name{getWBstats}
\alias{getWBstats}
\title{Calculate statistics for the first-tier mapping}
\usage{
getWBstats(seu, group.by)
}
\arguments{
\item{seu}{Modified query Seurat object from \code{mapToMB()}.}

\item{group.by}{Name of one metadata column to group cells by.}
}
\value{
A matrix of 37 rows. Each column corresponds to one group specified by group.by. 
Row 1-2 detail the proportion of confidently assigned and unassigned cells from the first-tier mapping
Row 3-11 detail the region.lineage proportions for confidently assigned cells
Row 12-34 detail the cell type proportions for confidently assigned cells
Row 35-37 detail the brain region proportions for confidently assigned DA neurons
}
\description{
Calculate statistics for the first-tier mapping
}
\author{
Lisheng Xu
}
