% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapToMB.r
\name{mapToMB}
\alias{mapToMB}
\title{Map query dataset to the fetal midbrain subatlas}
\usage{
mapToMB(seuMBRef, seuQuery, min.nCell = 1000, returnAnchor = FALSE)
}
\arguments{
\item{seuMBRef}{The Seurat object of the fetal midbrain subatlas, usually downloaded by function \code{downloadMBref()}}

\item{seuQuery}{The Seurat object of the query dataset, with log-normalised data and PC calculated, 
which has been mapped to the fetal whole brain atlas and contains predicted.reg.celltype.major in the metadata}

\item{min.nCell}{Minimal count of midbrain cells according to the first-tier mapping in seuQuery to allow it 
to be mapped by \code{mapToMB()}}

\item{returnAnchor}{Logical. Setting it to TRUE allows the transfer anchors to be returned}
}
\value{
A modified query Seurat object containing:
new assay corresponding to predicted.MB.celltype
an integrated reduction
a projected UMAP reduction 
new columns in the metadata: predicted.MB.celltype, predicted.MB.celltype.score and predicted.ventralScore. 
If returnAnchor is set to TRUE, a list containing such Seurat object and the transfer anchors is returned.
}
\description{
Map query dataset to the fetal midbrain subatlas
}
\author{
Lisheng Xu
}
