% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapToWB.r
\name{mapToWB}
\alias{mapToWB}
\title{Map query dataset to the fetal whole brain atlas}
\usage{
mapToWB(seuWBRef, seuQuery, returnAnchor = FALSE)
}
\arguments{
\item{seuWBRef}{The Seurat object of the fetal whole brain atlas, usually downloaded by function \code{downloadWBref()}}

\item{seuQuery}{The Seurat object of the query dataset, with log-normalised data and PC calculated}

\item{returnAnchor}{Logical. Setting it to TRUE allows the transfer anchors to be returned}
}
\value{
A modified query Seurat object containing:
new assay corresponding to predicted.celltype
an integrated reduction
a projected UMAP reduction 
new columns in the metadata: is.assigned, predicted.celltype, predicted.reg.celltype, predicted.reg.celltype.major and multiple scores. 
If returnAnchor is set to TRUE, a list containing such Seurat object and the transfer anchors is returned.
}
\description{
Map query dataset to the fetal whole brain atlas
}
\author{
Lisheng Xu
}
