% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMBstats.r
\name{plotMBstats}
\alias{plotMBstats}
\title{Plot statistics for the second-tier projection using \code{ComplexHeatmap}}
\usage{
plotMBstats(inpMat, no.legend = FALSE)
}
\arguments{
\item{inpMat}{A matrix from \code{getMBstats()}.}

\item{no.legend}{If true, a heatmap without legend is returned}
}
\value{
A heatmap constructed using \code{ComplexHeatmap}
}
\description{
Plot statistics for the second-tier projection using \code{ComplexHeatmap} (Fig 4)
}
\author{
Lisheng Xu
}
