% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWBstats.r
\name{plotWBstats}
\alias{plotWBstats}
\title{Plot statistics for the first-tier mapping using \code{ComplexHeatmap}}
\usage{
plotWBstats(inpMat, no.legend = FALSE)
}
\arguments{
\item{inpMat}{A matrix from \code{getWBstats()}.}

\item{no.legend}{If true, a heatmap without legend is returned}
}
\value{
A heatmap constructed using \code{ComplexHeatmap}
}
\description{
Plot statistics for the first-tier mapping using \code{ComplexHeatmap} (Fig 3)
}
\author{
Lisheng Xu
}
