% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWBstats.r
\name{plotWBstats}
\alias{plotWBstats}
\title{Plot statistics for the first-tier mapping using \code{ComplexHeatmap}}
\usage{
plotWBstats(
  inpMat,
  plot = c(1, 2, 3, 4),
  show.manno = FALSE,
  no.legend = FALSE
)
}
\arguments{
\item{inpMat}{A matrix from \code{getWBstats()}, or a list of multiple matrices. If names detected in the list, they are shown in the legend as study names.}

\item{plot}{blocks of the plots to be printed, i.e. c(1,2,3,4)
1: the proportion of confidently assigned and unassigned cells from the first-tier mapping
2: the region.lineage proportions for confidently assigned cells
3: the cell type proportions for confidently assigned cells
4: detail the brain region proportions for confidently assigned DA neurons}

\item{show.manno}{Display an in vivo fetal midbrain dataset published by La Manno 2016 for comparison.}

\item{no.legend}{If true, a heatmap without legend is returned}
}
\value{
A heatmap constructed using \code{ComplexHeatmap}
}
\description{
Plot statistics for the first-tier mapping using \code{ComplexHeatmap} (Fig 3)
}
\author{
Lisheng Xu
}
